+++ CHANGELOG.md
@@ -4,6 +4,8 @@
  - Switched to static linking for libgit2.
  - Added support for Qt 6.4.
  - Added file caching to improve speed.
+ - Fixed error when running multiple instances with the same url.
+ - Added a warning when a local path is chosen.
 
 ## Version 1.0.0 (August 2026)